requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
coloredlogs==15.0.1
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
dataclasses-json==0.5.14
EbookLib==0.18
effdet==0.4.1
emoji==2.8.0
et-xmlfile==1.1.0
faiss-cpu==1.7.4
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
langchain==0.0.271
langsmith==0.0.26
layoutparser==0.3.4
lit==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.15.1
openai==0.27.9
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==9.5.0
portalocker==2.7.0
protobuf==4.24.1
pycocotools==2.0.7
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pypandoc==1.11
pyparsing==3.0.9
pypdfium2==4.18.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
timm==0.9.5
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.32.0
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.10.5
unstructured-inference==0.5.16
urllib3==2.0.4
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
